src/main/java/com/epam/rd/autotasks/Sorting.java
package com.epam.rd.autotasks;

import java.util.Arrays;

public class Sorting {
    public void sort(int[] array){
        if(array==null){
            throw new IllegalArgumentException();
        }
        int n=array.length;
        for (int i=0;i<n-1;i++){
            for (int j=0;j<n-i-1;j++){
                if (array[j] > array[j + 1]) {
                    int temp = array[j];
                    array[j] = array[j + 1];
                    array[j + 1] = temp;
                }
            }
        }
//        Arrays.sort(array);
//        throw new UnsupportedOperationException();
    }
}
